requirements.txt
playwright==1.40.0
elevenlabs==0.2.27
moviepy==1.0.3
python-dotenv==1.0.0
requests==2.31.0
pytest-playwright
